Finish Ransom Note Hash/Dictionary Problem.
Be able to compare the words in two separate arrays by counting their
existence and be able to tell if one list can satisfy the other.

Not super tricky but, definitely got hung up when the complexity went
through the roof. Building a new hash where we can keep count of
previously seen strings and simply bump their occurrence count is easier
than building two separate arrays, one with the count and the other with
the string and then merging them with to_h. For example:

word_hash = magazine_words.to_h { |word| [word, magazine_words.count(word)] }

The performance of the above drops off considerably when not wanting to
assign or bump numbers with a ternary.

Hackerrank Url:
https://www.hackerrank.com/challenges/ctci-ransom-note