Finish Last Easy Hash Problem.
Rather simple (much easier than the previous one) problem of figuring
out if there is a char match from one string to another. There is
definitely a way of doing this using just chars but could be expensive
to scan an array if the strings themselves got super long.

Utilize the access pattern of hashes to break and set the result
variable to the correct/expected value. Included test

Hackerrank URL:
https://www.hackerrank.com/challenges/two-strings/