Available Calendar Slots.
Kind of a knapsack problem, kind of just a bizarre way to see how one
counts time. However, build a quick and dirty way to be able to take a
list of schedules and discern the open space or available space in them.
Utilize these spaces to recreate a schedule of open spaces.

Not exactly a HackerRank problem but, a problem brought to me by a
mentee during their job hunt. Honestly, I hadn't played with the
knapsack problem like this before and thought it would be a good
challenge.