Revert Back To Utilizing Math.
Math chosen for the first implementation and then ultimately abandoned
to explore building the entire array. However, it's just wildly silly to
do that when the list of space stations is usually much smaller.

Return to the math oriented approach and fix the issue with the first
and last space station and the first and last item in the array. This
passes all the HackerRank checks.